Fix bash command when initializing keycloakify project
In the bash command section, the remove yarn.lock file should be 

```bash
rm apps/keycloak-theme/yarn.lock
```

However the command is trying to delete `.yarn.lock` file. I assume this was a previous file made as hidden, but from the keycloakify-starter project, the `yarn.lock` file is not hidden.